Rename qiskit.execute module to qiskit.execute_function
The qiskit.execute module has a name conflict with the re-exported
path to it's sole public function qiskit.execute(). Whether python uses
the module or the function is an import time race condition and by
making the element imports occur at runtime this switches the normal
evaluation to use the module instead of the function. However, this was
always a bug and could have been triggered in other ways too. To make
the imports deterministic either the function or the module needs to be
renamed and since the module is less commonly used than the function
that is renamed.
