+++ docs/apidocs/execute.rst
@@ -1,6 +1,6 @@
 .. _qiskit-execute:
 
-.. automodule:: qiskit.execute
+.. automodule:: qiskit.execute_function
    :no-members:
    :no-inherited-members:
    :no-special-members:
