+++ test/python/circuit/test_parameters.py
@@ -28,7 +28,7 @@ from qiskit.circuit import (Gate, Instruction, Parameter, ParameterExpression,
                             ParameterVector)
 from qiskit.circuit.exceptions import CircuitError
 from qiskit.compiler import assemble, transpile
-from qiskit.execute import execute
+from qiskit.execute_function import execute
 from qiskit import pulse
 from qiskit.quantum_info import Operator
 from qiskit.test import QiskitTestCase
